refactor: list관련 button 공통 로직 분리
